+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 223:
 
 * Add test::stream::service
 * Add websocket service
+* Pausation abandoning test
 
 --------------------------------------------------------------------------------